Move more files for src/graphql (#36577)
