+++ src/graphql/components/Interface.tsx
