+++ src/graphql/components/Notice.tsx
